Closes issue #17. Implements advantage kit to intake
Fixed some variable names, updated he spin intake command, added the io layer and wrapper layer.
