+++ src/main/java/frc/robot/subsystems/Intake.java
@@ -1,25 +1,31 @@
+// Copyright (c) FIRST and other WPILib contributors.
+// Open Source Software; you can modify and/or share it under the terms of
+// the WPILib BSD license file in the root directory of this project.
+
 package frc.robot.subsystems;
 
-import com.revrobotics.CANSparkMax;
-import com.revrobotics.CANSparkMax.IdleMode;
-import com.revrobotics.CANSparkMaxLowLevel.MotorType;
-
 import edu.wpi.first.wpilibj2.command.SubsystemBase;
-public class Intake extends SubsystemBase{
-    private final CANSparkMax intakeMotor;
+import frc.robot.subsystems.IntakeIO.IntakeIOInputs;
 
-    public Intake(int intakeMotorID){
-        this.intakeMotor = new CANSparkMax(intakeMotorID, MotorType.kBrushless);
-        this.intakeMotor.setIdleMode(IdleMode.kBrake);
-        this.intakeMotor.setInverted(false);
-        this.intakeMotor.setSmartCurrentLimit(40);
-    }
+public class Intake extends SubsystemBase {
+  /** Creates a new Intake. */
+  private IntakeIO io;
+  private final IntakeIOInputsAutoLogged inputs = new IntakeIOInputsAutoLogged();
 
-    public void setIMotorSpeed(double intakeSpeed){
-        intakeMotor.set(intakeSpeed);
-    }
+  public Intake(IntakeIOInputs io) {
+    this.io = io;
+  }
 
-    public void stopMotor(){
-        intakeMotor.stopMotor();
-    }
+  public void setIntakeMotorSpeed(double intakeMotorSpeed) {
+    io.setIntakeMotorSpeed(intakeMotorSpeed);
+  }
+  
+  public void stopIntakeMotor() {
+    io.stopMotor();
+  }
+
+  @Override
+  public void periodic() {
+    // This method will be called once per scheduler run
+  }
 }
